Remove "What I'm Good At"
Selling yourself only goes so far, and that section was absurdly narcissistic. I wasn't even happy with the writing in it.

Correct spelling of "visualization" in DUG info.

Update Pivotal Labs info a bit more. Takes out all references to Xtreme (other than in the title) and more accurately reflects what we're doing as a company.

Expanded on the "tech I like" to mention more than just Java, and throw in an OwnCloud mention too. It's pretty cool, and I do indeed like it.